Reworking structure of City/CityParser classes
After having a long think, and several internal debates, I decided that parsing of a City object should be handled by the CityParser class (acting as a factory), especially considering the scope of this project.
Turned City's DataLength into a property instead of also having an interval variable
Modified City's constructor to account for the parser reorganization.